resources directory
