+++ .gitignore
@@ -1,10 +1,8 @@
 HELP.md
-src/main/resources/application.properties
 target/
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
 !**/src/test/**/target/
-**/src/main/**/resources/application.properties
 
 ### STS ###
 .apt_generated
