+++ src/main/java/com/sparta/eng80/onetoonetracker/OneToOneTrackerApplication.java
@@ -1,0 +1,13 @@
+package com.sparta.eng80.onetoonetracker;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OneToOneTrackerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(OneToOneTrackerApplication.class, args);
+    }
+
+}
