+++ src/main/java/com/sparta/eng80/onetoonetracker/entities/FeedbackEntity.java
@@ -1,0 +1,4 @@
+package com.sparta.eng80.onetoonetracker.entities;
+
+public class FeedbackEntity {
+}
