+++ src/main/java/com/sparta/eng80/onetoonetracker/entities/GroupEntity.java
@@ -1,0 +1,4 @@
+package com.sparta.eng80.onetoonetracker.entities;
+
+public class GroupEntity {
+}
