+++ src/main/java/com/sparta/eng80/onetoonetracker/entities/UserEntity.java
@@ -1,0 +1,4 @@
+package com.sparta.eng80.onetoonetracker.entities;
+
+public class UserEntity {
+}
