+++ src/main/java/com/sparta/eng80/onetoonetracker/repositories/TraineeRepository.java
@@ -1,0 +1,9 @@
+package com.sparta.eng80.onetoonetracker.repositories;
+
+import com.sparta.eng80.onetoonetracker.entities.TraineeEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TraineeRepository extends CrudRepository<TraineeEntity, Integer> {
+}
