+++ src/main/java/com/sparta/eng80/onetoonetracker/repositories/TrainerRepository.java
@@ -1,0 +1,8 @@
+package com.sparta.eng80.onetoonetracker.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TrainerRepository extends CrudRepository<TrainerRepository, Integer> {
+}
