+++ src/main/java/com/sparta/eng80/onetoonetracker/repositories/UserRepository.java
@@ -1,0 +1,9 @@
+package com.sparta.eng80.onetoonetracker.repositories;
+
+import com.sparta.eng80.onetoonetracker.entities.UserEntity;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<UserEntity, Integer> {
+}
