+++ src/main/java/com/sparta/eng80/onetoonetracker/services/LoginCredentialService.java
@@ -1,0 +1,15 @@
+package com.sparta.eng80.onetoonetracker.services;
+
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class LoginCredentialService implements UserDetailsService {
+
+    @Override
+    public UserDetails loadUserByUsername(String s) throws UsernameNotFoundException {
+        return null;
+    }
+}
