+++ src/main/java/com/sparta/eng80/onetoonetracker/services/interfaces/AdminAppService.java
@@ -1,0 +1,9 @@
+package com.sparta.eng80.onetoonetracker.services.interfaces;
+
+import com.sparta.eng80.onetoonetracker.entities.TrainerEntity;
+
+public interface AdminAppService {
+
+    TrainerEntity saveTrainer(TrainerEntity trainer);
+    TrainerEntity deleteTrainer(TrainerEntity trainer);
+}
