+++ src/main/java/com/sparta/eng80/onetoonetracker/services/interfaces/ApplicationService.java
@@ -1,0 +1,10 @@
+package com.sparta.eng80.onetoonetracker.services.interfaces;
+
+import java.util.Optional;
+
+public interface ApplicationService<T> {
+
+    Optional<T> findById();
+    Iterable<T> findAll();
+    T save(T t);
+}
