+++ src/main/java/com/sparta/eng80/onetoonetracker/services/interfaces/FeedbackAppService.java
@@ -1,0 +1,10 @@
+package com.sparta.eng80.onetoonetracker.services.interfaces;
+
+import com.sparta.eng80.onetoonetracker.entities.FeedbackEntity;
+import com.sparta.eng80.onetoonetracker.entities.TrainerEntity;
+
+public interface FeedbackAppService extends ApplicationService<FeedbackEntity> {
+
+    Iterable<FeedbackEntity> findByTrainee(TrainerEntity trainee);
+    Iterable<FeedbackEntity> findByTrainer(TrainerEntity trainer);
+}
