+++ src/main/java/com/sparta/eng80/onetoonetracker/services/interfaces/GroupAppService.java
@@ -1,0 +1,9 @@
+package com.sparta.eng80.onetoonetracker.services.interfaces;
+
+import com.sparta.eng80.onetoonetracker.entities.GroupEntity;
+
+public interface GroupAppService extends ApplicationService<GroupEntity> {
+
+    GroupEntity findByName(String name);
+    GroupEntity findByStreamId(int id);
+}
