--- src/main/java/com/sparta/eng80/spartaonetooneformtracker/SpartaOneToOneFormTrackerApplication.java
@@ -1,13 +1,0 @@
-package com.sparta.eng80.spartaonetooneformtracker;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpartaOneToOneFormTrackerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SpartaOneToOneFormTrackerApplication.class, args);
-    }
-
-}
