--- src/main/java/com/sparta/eng80/spartaonetooneformtracker/entities/TraineeEntity.java
@@ -1,4 +1,0 @@
-package com.sparta.eng80.spartaonetooneformtracker.entities;
-
-public class TraineeEntity {
-}
