--- src/main/java/com/sparta/eng80/spartaonetooneformtracker/entities/UserEntity.java
@@ -1,4 +1,0 @@
-package com.sparta.eng80.spartaonetooneformtracker.entities;
-
-public class UserEntity {
-}
