--- src/main/java/com/sparta/eng80/spartaonetooneformtracker/repositories/TraineeRepository.java
@@ -1,9 +1,0 @@
-package com.sparta.eng80.spartaonetooneformtracker.repositories;
-
-import com.sparta.eng80.spartaonetooneformtracker.entities.TraineeEntity;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TraineeRepository extends CrudRepository<TraineeEntity, Integer> {
-}
