--- src/main/java/com/sparta/eng80/spartaonetooneformtracker/services/TrainerService.java
@@ -1,56 +1,0 @@
-package com.sparta.eng80.spartaonetooneformtracker.services;
-
-import com.sparta.eng80.spartaonetooneformtracker.entities.TrainerEntity;
-import com.sparta.eng80.spartaonetooneformtracker.services.interfaces.UserAppService;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-public class TrainerService implements UserAppService<TrainerEntity> {
-
-    @Override
-    public Optional<TrainerEntity> findById() {
-        return Optional.empty();
-    }
-
-    @Override
-    public Iterable<TrainerEntity> findAll() {
-        return null;
-    }
-
-    @Override
-    public TrainerEntity save(TrainerEntity trainerEntity) {
-        return null;
-    }
-
-    @Override
-    public Optional<TrainerEntity> findByUserId(int id) {
-        return Optional.empty();
-    }
-
-    @Override
-    public Optional<TrainerEntity> findByEmail(String name) {
-        return Optional.empty();
-    }
-
-    @Override
-    public Iterable<TrainerEntity> findByGroupId(int id) {
-        return null;
-    }
-
-    @Override
-    public Iterable<TrainerEntity> findByFirstName(String name) {
-        return null;
-    }
-
-    @Override
-    public Iterable<TrainerEntity> findByLastName(String name) {
-        return null;
-    }
-
-    @Override
-    public Iterable<TrainerEntity> findByName(String first, String last) {
-        return null;
-    }
-}
