--- src/main/java/com/sparta/eng80/spartaonetooneformtracker/services/interfaces/AdminAppService.java
@@ -1,9 +1,0 @@
-package com.sparta.eng80.spartaonetooneformtracker.services.interfaces;
-
-import com.sparta.eng80.spartaonetooneformtracker.entities.TrainerEntity;
-
-public interface AdminAppService {
-
-    TrainerEntity saveTrainer(TrainerEntity trainer);
-    TrainerEntity deleteTrainer(TrainerEntity trainer);
-}
