--- src/main/java/com/sparta/eng80/spartaonetooneformtracker/services/interfaces/ApplicationService.java
@@ -1,10 +1,0 @@
-package com.sparta.eng80.spartaonetooneformtracker.services.interfaces;
-
-import java.util.Optional;
-
-public interface ApplicationService<T> {
-
-    Optional<T> findById();
-    Iterable<T> findAll();
-    T save(T t);
-}
