--- src/test/java/com/sparta/eng80/spartaonetooneformtracker/SpartaOneToOneFormTrackerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sparta.eng80.spartaonetooneformtracker;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpartaOneToOneFormTrackerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
